requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# GPIO Control (Raspberry Pi)
RPi.GPIO==0.7.1

# Serial Communication (DMX)
pyserial==3.5

# Audio/Sound Support
pygame==2.5.2

# Additional utilities
python-json-logger==2.0.7
watchdog==3.0.0

# Development dependencies (optional)
# flask-cors==4.0.0
# gunicorn==21.2.0
